Move StoreTrainingData to training_wrapper_functions.py
